ext/zend_test: Move object handler test objects to their own file
