+++ ext/zend_test/config.m4
@@ -4,5 +4,5 @@ PHP_ARG_ENABLE([zend-test],
     [Enable zend_test extension])])
 
 if test "$PHP_ZEND_TEST" != "no"; then
-  PHP_NEW_EXTENSION(zend_test, test.c observer.c fiber.c, $ext_shared,, -DZEND_ENABLE_STATIC_TSRMLS_CACHE=1)
+  PHP_NEW_EXTENSION(zend_test, test.c observer.c fiber.c object_handlers.c, $ext_shared,, -DZEND_ENABLE_STATIC_TSRMLS_CACHE=1)
 fi
