+++ ext/zend_test/object_handlers.h
@@ -1,0 +1,22 @@
+/*
+  +----------------------------------------------------------------------+
+  | Copyright (c) The PHP Group                                          |
+  +----------------------------------------------------------------------+
+  | This source file is subject to version 3.01 of the PHP license,      |
+  | that is bundled with this package in the file LICENSE, and is        |
+  | available through the world-wide-web at the following url:           |
+  | https://www.php.net/license/3_01.txt                                 |
+  | If you did not receive a copy of the PHP license and are unable to   |
+  | obtain it through the world-wide-web, please send a note to          |
+  | license@php.net so we can mail you a copy immediately.               |
+  +----------------------------------------------------------------------+
+  | Authors: George Peter Banyard <girgias@php.net>                        |
+  +----------------------------------------------------------------------+
+*/
+
+#ifndef ZEND_TEST_OBJECT_HANDLERS_H
+#define ZEND_TEST_OBJECT_HANDLERS_H
+
+void zend_test_object_handlers_init(void);
+
+#endif
